src/worker/requirements.txt
pika==1.3.2
python-decouple==3.8
pyTelegramBotAPI~=4.15.4
aiohttp~=3.9.3